% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_preprocessing.R
\name{backend_preprocessing}
\alias{backend_preprocessing}
\title{backend_preprocessing}
\usage{
backend_preprocessing(datatable, target_col, task_type, positive = NULL)
}
\arguments{
\item{datatable}{A data.table object, holding the target variable and
features.}

\item{target_col}{A character. The name of the target variable.}

\item{task_type}{A character. The type of learning task to prepare
`datatable` for. Can be one of `regression`, `class:binary` or
`class:multiclass`.}

\item{positive}{A character. If `task_type` = `class:binary`, this argument
is required to specify the positive class of the binary classification
task, which will be replaced by the value `1` in the resulting object.}
}
\value{
The function returns a data.table with the transformed target
  variable and feature variables. This object can then be used to create
  an mlr3 task to be used with the \code{mlr3learners.lightgbm} R package.
}
\description{
The function transforms the values of a data.table object
  so that in can be passed to an mlr3 task to be used with the
  \code{mlr3learners.lightgbm} R package. This is necessary because
  \code{lightgbm} can only handle numeric and integer features. Furthermore,
  for classification tasks the target variable needs to be of type integer
  with values starting at '0' for the first class.
  The function is a wrapper around \code{lightgbm::prepare}.
}
\seealso{
\code{mlr3}, \code{lightgbm::prepare}, \code{mlr3learners.lightgbm},
  \code{plyr::revalue}
}
