% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_plot.R
\name{importance_plot}
\alias{importance_plot}
\title{importance_plot}
\usage{
importance_plot(importance, n = NULL, threshold = NULL)
}
\arguments{
\item{importance}{A named vector of importance values. If supported,
the output of \code{learner$importance}.}

\item{n}{An integer. The maximum number of importance values included
into the plot (default: NULL, indicating all importance values will
be plotted). Restrictions: 0 < n <= `length(importance)`.}

\item{threshold}{A numeric. The threshold of importance values to be
included into the plot (default: NULL). Restrictions: 0 < threshold < 1.
Please note, that if `n` and `threshold` are provided, `threshold` is
ignored.}
}
\value{
An importance plot.
}
\description{
The function returns a barplot for feature importance vectors.
}
