% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrLightGBM.R
\name{LearnerRegrLightGBM}
\alias{LearnerRegrLightGBM}
\alias{mlr_learners_regr.lightgbm}
\title{Regression LightGBM Learner}
\format{[R6::R6Class] inheriting from [mlr3::LearnerRegr].}
\description{
Regression LightGBM Learner

Regression LightGBM Learner
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3:LearnerRegr]{mlr3::LearnerRegr}} -> \code{LearnerRegrLightGBM}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nrounds}}{Number of training rounds.}

\item{\code{early_stopping_rounds}}{A integer. Activates early stopping.
Requires at least one validation data and one metric. If there's
more than one, will check all of them except the training data.
Returns the model with (best_iter + early_stopping_rounds).
If early stopping occurs, the model will have 'best_iter' field.}

\item{\code{categorical_feature}}{A list of str or int. Type int represents
index, type str represents feature names.}

\item{\code{train_data}}{A data.table object holding the training data.}

\item{\code{train_label}}{A vector holding the training labels.}

\item{\code{valid_data}}{A data.table object holding the validation data.}

\item{\code{valid_label}}{A vector holding the validation labels.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LearnerRegrLightGBM$new()}}
\item \href{#method-train_internal}{\code{LearnerRegrLightGBM$train_internal()}}
\item \href{#method-predict_internal}{\code{LearnerRegrLightGBM$predict_internal()}}
\item \href{#method-importance}{\code{LearnerRegrLightGBM$importance()}}
\item \href{#method-importance2}{\code{LearnerRegrLightGBM$importance2()}}
\item \href{#method-valids}{\code{LearnerRegrLightGBM$valids()}}
\item \href{#method-clone}{\code{LearnerRegrLightGBM$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format">}\href{../../mlr3/html/Learner.html#method-format}{\code{mlr3::Learner$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help">}\href{../../mlr3/html/Learner.html#method-help}{\code{mlr3::Learner$help()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict">}\href{../../mlr3/html/Learner.html#method-predict}{\code{mlr3::Learner$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata">}\href{../../mlr3/html/Learner.html#method-predict_newdata}{\code{mlr3::Learner$predict_newdata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print">}\href{../../mlr3/html/Learner.html#method-print}{\code{mlr3::Learner$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset">}\href{../../mlr3/html/Learner.html#method-reset}{\code{mlr3::Learner$reset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train">}\href{../../mlr3/html/Learner.html#method-train}{\code{mlr3::Learner$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
The initialize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_internal"></a>}}
\subsection{Method \code{train_internal()}}{
The train_internal function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$train_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_internal"></a>}}
\subsection{Method \code{predict_internal()}}{
The predict_internal function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$predict_internal(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance"></a>}}
\subsection{Method \code{importance()}}{
The importance function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$importance()}\if{html}{\out{</div>}}
}

\subsection{Details}{
A named vector with the learner's variable importances.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance2"></a>}}
\subsection{Method \code{importance2()}}{
The importance2 function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$importance2()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a list with the learner's variable importance values
  and an importance plot.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valids"></a>}}
\subsection{Method \code{valids()}}{
The valids function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$valids(task, row_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{The mlr3 task.}

\item{\code{row_ids}}{An integer vector with the row IDs for the validation
data.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function can be used to provide a subsample to the data
  to the lightgbm's train function's `valids` argument. This is e.g.
  needed, when the argument `early_stopping_rounds` is used.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerRegrLightGBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
