% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_prauc.R
\name{lgb_prauc}
\alias{lgb_prauc}
\title{lgb_rmsle custom metric}
\usage{
lgb_prauc(preds, dtrain)
}
\arguments{
\item{preds}{A vector containing the predicted values}

\item{dtrain}{The training dataset.}
}
\description{
PR-AUC custom evaluation metric function
}
