% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_rmsle.R
\name{lgb_rmsle}
\alias{lgb_rmsle}
\title{lgb_rmsle custom metric}
\usage{
lgb_rmsle(preds, dtrain)
}
\arguments{
\item{preds}{A vector containing the predicted values}

\item{dtrain}{The training dataset.}
}
\description{
RMSLE custom evaluation metric function
}
