% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightGBM.R
\name{LightGBM}
\alias{LightGBM}
\title{LightGBM Learner}
\description{
LightGBM Learner

LightGBM Learner
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{nrounds}}{Number of training rounds.}

\item{\code{early_stopping_rounds}}{A integer. Activates early stopping.
Requires at least one validation data and one metric. If there's
more than one, will check all of them except the training data.
Returns the model with (best_iter + early_stopping_rounds).
If early stopping occurs, the model will have 'best_iter' field.}

\item{\code{categorical_feature}}{A list of str or int. Type int represents
index, type str represents feature names.}

\item{\code{train_data}}{A data.table object holding the training data.}

\item{\code{train_label}}{A vector holding the training labels.}

\item{\code{valid_data}}{A data.table object holding the validation data.}

\item{\code{valid_label}}{A vector holding the validation labels.}

\item{\code{label_names}}{The unique label names in classification tasks.}

\item{\code{trans_tar}}{The transfrom_target instance.}

\item{\code{param_set}}{The lightgbm parameters.}

\item{\code{nrounds_by_cv}}{A logical. Calculate the best nrounds by using
the `lgb.cv` before the training step}

\item{\code{cv_folds}}{The number of cross validation folds, when setting
`nrounds_by_cv` = TRUE (default: 5).}

\item{\code{cv_model}}{The cross validation model.}

\item{\code{model}}{The trained lightgbm model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LightGBM$new()}}
\item \href{#method-train_cv}{\code{LightGBM$train_cv()}}
\item \href{#method-train}{\code{LightGBM$train()}}
\item \href{#method-predict}{\code{LightGBM$predict()}}
\item \href{#method-importance}{\code{LightGBM$importance()}}
\item \href{#method-importance2}{\code{LightGBM$importance2()}}
\item \href{#method-valids}{\code{LightGBM$valids()}}
\item \href{#method-clone}{\code{LightGBM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
The data_preprocessing function


The initialize function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$new()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train_cv"></a>}}
\subsection{Method \code{train_cv()}}{
The train_cv function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$train_cv(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\subsection{Method \code{train()}}{
The train function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$train(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\subsection{Method \code{predict()}}{
The predict function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$predict(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{An mlr3 task}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance"></a>}}
\subsection{Method \code{importance()}}{
The importance function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$importance()}\if{html}{\out{</div>}}
}

\subsection{Details}{
A named vector with the learner's variable importances.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-importance2"></a>}}
\subsection{Method \code{importance2()}}{
The importance2 function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$importance2()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns a list with the learner's variable importance values
  and an importance plot.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-valids"></a>}}
\subsection{Method \code{valids()}}{
The valids function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$valids(task, row_ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{The mlr3 task.}

\item{\code{row_ids}}{An integer vector with the row IDs for the validation
data.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The function can be used to provide a subsample to the data
  to the lightgbm's train function's `valids` argument. This is e.g.
  needed, when the argument `early_stopping_rounds` is used.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LightGBM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
