% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.lightgbm-package}
\alias{mlr3learners.lightgbm}
\alias{mlr3learners.lightgbm-package}
\title{mlr3learners.lightgbm: Learners from lightgbm package for mlr3}
\description{
Adds `lgb.train()` from the lightgbm package to
    mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.lightgbm}
  \item Report bugs at \url{https://github.com/mlr3learners/mlr3learners.lightgbm/issues}
}

}
\author{
\strong{Maintainer}: Lorenz A. Kapsner \email{lorenz.kapsner@gmail.com} (\href{https://orcid.org/0000-0003-1866-860X}{ORCID})

}
